+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 * `Versions`:
   * `MicroUtils`: `0.16.2` -> `0.16.4`
 * `Core`:
-    * Simplify default `RequestsLimiter` (`ExceptionsOnlyLimiter`)
+    * Simplify default `RequestsLimiter` (`ExceptionsOnlyLimiter`) (thanks to @y9san9 for help)
 
 ## 4.2.2