+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # TelegramBotAPI changelog
 
+## 4.2.3
+
+* `Versions`:
+  * `MicroUtils`: `0.16.2` -> `0.16.4`
+* `Core`:
+    * Simplify default `RequestsLimiter` (`ExceptionsOnlyLimiter`) (thanks to [@y9san9](https://github.com/y9san9) for help)
+
 ## 4.2.2
 
 * `Versions`: